Added Signature Commercial

+++ Options/SignatureCommercialbuildingOptions.cs
@@ -5,10 +5,10 @@ public class SignatureCommercialbuildingOptions
     public string Name { get; set; } = null!;
 
     public int Attractiveness { get; set; }
-    public int Wellbeing { get; set; }
-    public int Radius { get; set; }
-    public int CityImportcost { get; set; }
-    public int CityAttractiveness { get; set; }
+    public float Wellbeing { get; set; }
+    public float Radius { get; set; }
+    public float CityImportcost { get; set; }
+    public float CityAttractiveness { get; set; }
     public int Noisepollution { get; set; }
     
 }
